kernel: exportfs: reallow building as a module
It was originally tristate, but was changed to bool to prevent a build
breakage in an earlier patch revision (for a new syscall). The actual
commit[1] changed it in a way this isn't needed anymore, but
accidentially left the EXPORTFS change in. There's no reason for it
not being buildable as a module, so changing it back to tristate is
fine.

see: https://patchwork.kernel.org/patch/812832/

Thank you Jonas Gorski for the patch.